A PHP and HTML based web application which mimics an e-commerce website.
This web application allows the user to brwose products of Compares,
which is a local chocolate shop in LA, insert, select,  delete, and
update information about chocolate bars in an existing database. All of
the code are valid; however, the database I used through UMUC is closed
after the class. Therefore, when the user accesses the database, a
webpage for displaying the results from the database is broken.